SEBSERV-877 use SEBSettingsPanel for  both, SEBSettingsDialog and SEBSettings(Step)
